Update `CorsetValidator` with Expansion Options
This updates the `CorsetValidator` to support various configuration
options used by `corset`.  Specifically, for field arithmetic and
constraint expansion.  A better set of defaults is used which means
testing now reflects more closely what the prover sees.  However, we
stop short of going all the way to the lowest level because this has a
large impact on performance.  Nevertheless, the option to do this is
present in the updated `CorsetValidator` and can be enabled easily
enough.