Configure `CorsetValidator` via `CORSET_FLAGS`
This allows the `CorsetValidator` to be configured via an environment
variable `CORSET_FLAGS`.  This allows, for example, different levels of
expansion to be specified on the CI pipeline, or when running locally.

In addition, this updates the `CorsetValidator` to allow different
amounts of expansion to be specified.  Previously, it was "all or
nothing".  Now, upto four levels of expansion can be specified.