Regenerate `Trace.java` for all modules except hub
This regenerates the `Trace.java` for all modules except the hub.  This
enables more efficient `Trace.java` files to be created using the latest
version of `corset`.  The hub is not changed for now, simply because
there is a lot going on with this module (i.e. in other branches).
However, it can be updated as and when required.