Initial Integration of `go-corset`
This integrates `go-corset` and (when installed) runs it side-by-side
the existing Rust tool.  However, if the `go-corset` binary cannot be
found then it is simply ignored.

At this stage, `go-corset` is not enabled in the CI pipeline.  This is
because it is consuming too much memory on the test `leoFailingRange`.