Enable `go-corset` in `gradle` action
This enables `go-corset` in the `gradle` action workflow.  However,
`go-corset` is only run as part of a job which runs separately from the
main tests job.  As such, it will not prevent pull requests from being
merged.  The goal here is really to enable testing of `go-corset`
before (eventually) swapping out the existing Rust corset tool.