fix: Use correct CHAINID in EVM
This updates the `ToyExecutionEnvironment` to use the correct CHAINID
when constructing the EVM.  Previously, this was not done correctly and
meant that the `CHAINID` bytecode was returning an invalid value, which
in turn mean that storage addresses were being computed incorrectly
(i.e. because they incorporated the CHAINID into the hash).