Enable capture & checking of transaction outcomes
This updates the `BlockCapturer` to capture details related to
transaction outcomes.  Specifically, it records whether or not the
transaction suceeded, what output data was generated, what logs were
generated, how much gas was used and what accounts were self destructed.

The expected outcome is checked against the actual outcome, and a
suitable exception raised if there is a discrepency.  However, if the
outcome is missing (i.e. because its an old trace which pre-dates this
feature) then a log is reported that the outcome "was not checked".

Finally, it should be noted that we could attempt to capture more
information about what happened during transaction executing.
Specifically, what the state differences were.