Record touched state in `TransactionResultSnapshot`
This updates the `BlockCapturer` to record the set of accounts and
storage locations which were touched by a given transaction.  This
includes those which were modified, but also those which were just read.

This allows the testing mechanism to ensure that the state is exactly as
we expected after each transaction.