Refactor for transactions which are non-starters
This puts in place a refactoring which allows `txStorage` and
`txAddresses` to be initialised with `null`.  A key challenge was that
not all transactions go through `traceStartTransaction()`.  Instead,
some transactions go `tracePrepareTransaction()` and then straight to
`traceEndTransaction()`.

In addition, this refactors `TransactionChecker` so that it no longer
wraps an existing tracer.  Instead, we can "sequence" tracers using
`ConflationAwareOperationTracer.sequence()` making it relatively easy to
combine two tracers together.  Furthermore, this helps with testing the
`BlockCapturer` as it becomes easier to run it from
`ToyExecutionEnvironment`.