feat: update blockcapturer to capture outcomes (#982)
* Enable capture & checking of transaction outcomes

This updates the `BlockCapturer` to capture details related to
transaction outcomes.  Specifically, it records whether or not the
transaction suceeded, what output data was generated, what logs were
generated, how much gas was used and what accounts were self destructed.

The expected outcome is checked against the actual outcome, and a
suitable exception raised if there is a discrepency.  However, if the
outcome is missing (i.e. because its an old trace which pre-dates this
feature) then a log is reported that the outcome "was not checked".

Finally, it should be noted that we could attempt to capture more
information about what happened during transaction executing.
Specifically, what the state differences were.

* Record touched state in `TransactionResultSnapshot`

This updates the `BlockCapturer` to record the set of accounts and
storage locations which were touched by a given transaction.  This
includes those which were modified, but also those which were just read.

This allows the testing mechanism to ensure that the state is exactly as
we expected after each transaction.

* Fix Transaction Result Checking

There was a problem with the position at which result checking was being
performed, in that it didn't exactly match the position where the result
snapshot data was being generated.  This is resolved using a wrapper
around ZkTracer which handles checking of transaction results at the
correct position.

* Fix `ToyWorld.commit()`

This puts through a simple fix for `ToyWorld.commit()` which became
evident from block `2247098`.  Since this matches the code in
`SimpleWorld`, it seems to me there is a bug there.

* Refactor for transactions which are non-starters

This puts in place a refactoring which allows `txStorage` and
`txAddresses` to be initialised with `null`.  A key challenge was that
not all transactions go through `traceStartTransaction()`.  Instead,
some transactions go `tracePrepareTransaction()` and then straight to
`traceEndTransaction()`.

In addition, this refactors `TransactionChecker` so that it no longer
wraps an existing tracer.  Instead, we can "sequence" tracers using
`ConflationAwareOperationTracer.sequence()` making it relatively easy to
combine two tracers together.  Furthermore, this helps with testing the
`BlockCapturer` as it becomes easier to run it from
`ToyExecutionEnvironment`.

* Drop `Reaper.traceStartTransaction()`.