Capture block hashes in `BlockCapturer` for replay
This updates the `BlockCapturer` to capture block hashes seen during the
execution of a given conflation.  Such hashes are then stored in the
replay file such that they can be fed back into the EVM during replay.
Thus, the `BLOCKHASH` bytecode can determine what to return.

Two test cases are added for blocks `8718090` and `8718330` which are
known to include tranactions that execute BLOCKHASH.  These have been
seeded with block hash information so that they execute correctly.

NOTE: if any block hashes are missing from the replay file then an
exception is raised.  This is simply to ensure there is no spooky action
at a distance.