Expose true output from tx result failure
This updates the `TransactionResultChecker` to work around the fact that
the `MainnetTransactionProcessor` calls `traceEndTransaction()` twice
when a `RuntimeException` is raised.  Previously, this was causing the
true reason for a result failure.  Now, we can see the true reason.

Furthermore, this updates all the replay files to include tx outcomes,
and blockhashes (where appropriate).