Use `ReferenceTestWorldState` instead of `ToyWorld` (#1045)
This puts in place a temporary fix which converts `ToyWorld` over into
`ReferenceTestWorldState` to use when executing transactions.  The
conversion could be done more elegantly, but the goal here is to get
something out for people to use ASAP.  Eventually, we would then remove
`ToyWorld` altogether.

This also updates the replay files for various tests so they include the
updated block capture (e.g. with outcomes).  All tests which pass
outcome checking are marked as such, but there are a few remaining which
fail (e.g. because of incorrect balance or, in some cases, nonce).