feat: update replay tests with tx outcomes (#1091)
* Initial updates to replay files

This updates all replay files with outcome information.  This also
enables result checking on a reasonable number of tests, but there are
actually a lot which fail still for various reasons.  Furthermore, this
improves debugging information for `go-corset`.

* Tweak Memory Settings for Go Corset

This tweaks the memory settings for `go-corset` to reduce overall memory
pressure.