Tidy up `ToyExecutionEnvironment`
This tidies up the `ToyExecutionEnvironment` in several ways.  Firstly,
it updates the manner in which the `MainnetTransactionProcessor` is
created by reusing as much from BESU as possible; secondly, it removes
`ToyWorld` altogether (as this is no longer needed); finally, it removes
`ToyExecutionEnvironment.execute()` since this is no longer being used.