feat: tidy up `ToyExecutionEnvironment` and remove `ToyWorld` (#1143)
* Tidy up `ToyExecutionEnvironment`

This tidies up the `ToyExecutionEnvironment` in several ways.  Firstly,
it updates the manner in which the `MainnetTransactionProcessor` is
created by reusing as much from BESU as possible; secondly, it removes
`ToyWorld` altogether (as this is no longer needed); finally, it removes
`ToyExecutionEnvironment.execute()` since this is no longer being used.

* Support Mainnet vs Sepolia Tests

This adds support for distinguishing between Mainnet and Sepolia tests.
This is helpful for the odds cases where we want a replay which was
inteded for sepolia.

* Remove `getMainnetTransactionProcessorOrig()`

This removes the now redundant (original) method for getting the mainnet
transaction processor.

* Apply spotless

* Fix broken replay test.

`modexpTriggeringNonAlignedFirstLimbSingleSourceMmuModexp` was missing
the chain identifier.  I'm assuming LINEA_MAINNET for now.

---------

Co-authored-by: Olivier Bégassat <38285177+OlivierBBB@users.noreply.github.com>