Fix BlockCapturer for Created Accounts
The block capturer had an issue capturing accounts which were created
during the execution of the conflation being captured.  The problematic
scenario required an account to exist at the beginning of the
conflation, but which was then the subject of a CREATE/CREATE2
contract creation.  This is permitted in certain situations, such as
when the account only has a balance (but no storage).

To resolve this case, we have to be a little more careful when creating
storage snapshots.  Specifically, when an account only partially exists
(i.e. has no storage, but only a balance) then we never create storage
snapshots.

Finally, this also adds a debugging mode to ReplayExecutionEnvironment
to assist debugging the BlockCapturer.