Fix for broken replay files
This puts in place a simple fix for replay files which incorrectly
include StorageSnapshots for accounts which are to be created.  Such
snapshots confuse BESU into thinking their storage is not empty which,
in turn, prevents the accounts from being created (and then fails result
checking because the txs in question fail).

In principle, this fix should not be necessary because the BlockCapturer
has already been fixed to resolve this.  However, we still have a number
of older replay files in the repository and this fix resolves all of
time right now (i.e. so we don't have to wait for them to be recaptured,
etc).