Support regeneration of `GlobalConstants.java` (#1421)
This enables regeneration of the `GlobalConstants.java` via the command
`gradle constants`.  This requires Corset `v9.7.15` (which at this time
has not yet been released).  The solution is to allow for an override to
the class name being produced in Corset.  So, it doesn't always have to
generate a file `Trace.java`.  Rather you can specify the class name to
generate (in this case, `GlobalConstants`).  The relevant gradle option
for the `constants` command is updated to do this.