fix: added initializationCodes HashMap to DeploymentInfo
rationale: canonical account snapshots don't properly work on accounts
currently undergoing deployment. Indeed the code is viewed as empty.

This is corrected in this commit, in particular in how we associate code
to a snapshot depending on whether the code is currently underoing
deployment or not, i.e. depending on DEPLOYMENT_STATUS.

This fixes

jump-instruction---miscellaneous-row---setting-the-OOB-instruction---JUMP-case