Support Variable Report Width for Corset
This updates the `CorsetValidator` to support changing the number of
trace rows reported for a failing constraint.  The default is firstly
changed to `8` (compared with `2` previously).  Secondly, its possible
to override the default via the `CORSET_FLAGS` environment variable.
This means we can set a small width for the CI pipeline, where
horizontal real estate is more limited.