Disable implicit parallel forks
This disables the use of parallel forks in gradle unless this is
explicitly requested.  The issue is that parallel forks interferes with
go-corset which is highly concurrent.  Hence, running go-corset tasks in
parallel can (in some cases) lead to out-of-resource issues and workflow
tasks being cancelled unexpectedly.

However, since using parallel forks does improve the runtime for Rust
corset, this option is retained when specifically specified via the
REPLAY/WEEKLY/REFERENCE_TESTS_PARALLELISM environment variables.