Split Unit and Replay Tests (#1593)
* Split Unit and Replay Tests

The goal here is to split out the unit and replay tests into distinct
workflows.  This will allow them to be potentially executed in parallel,
leading to reduced time to PR.  It will also give slightly more feedback
in terms of which parts are failing.

* Support fine-grained parallelism for unit tests

This enables fine-grained parallelism for the unit tests, along with
some suitable defaults.  Specifically, this sets defaults for running on
a local machine assumed to have 8 cores.

* Update configuration for self-hosted runners

This updates the new configuration to make use of self-hosted runners.

* Refactor Unit / Replay Workflows

This refactors the workflows to keep them in one file.  In order to make
that work, have factored out some of the common components into a
separate "composite" action.

* Download Corset by default

This configures the CI pipeline to download corset by default, rather
than rebuilding it from scratch everytime.  However, it is still
possible to request that it be rebuilt from scratch (if/when that is
desirable).