fix: correct MARKED_FOR_SELFDESTRUCT
There is a subtle point with selt-targeting SELFDESTRUCT's: for the two
"doing" account-rows the HUB_STAMP remains the same. We therefore cannot
use the following definition

  markedForSelfDestruct = hubStamp > selfDestructTime

since the second time we inspect the account, this time as the
"recipient", the flag ought to be true. Instead we must differentiate
according to the DOM_STAMP which reads

- row 1: hubStamp * MULTIPLIER___DOM_SUB_STAMPS + 0
- row 2: hubStamp * MULTIPLIER___DOM_SUB_STAMPS + 1

This is what we fix here