Tweak memory settings
This tweaks the memory settings to facilitate better performance on the
CI server.  In fact, it lowers the amount of memory available.  This is
because we don't want Java and Go competing for the same amount of
memory.  Furthermore, we now know that all of the tests can be completed
on a machine with just 32Gb RAM.