fix: reworking of the TxSkipSection
We now extract the coinbase address from a MessageFrame. This ought to
fix an issue with replay tests with a TX_SKIP type of transaction.
Indeed, for some reason (maybe related to using clique and this changing
the "miner" field of the block header) running

  block.getCoinbaseAddress()

or so returned 0x00 rather than the expected 0x8f8.

Also greatly simplified this TraceSection.