Enable conditional build
This configures the build process to be conditional on a system-wide
property called "blockchain".  This property can be specified on the
command-line with e.g. "-Dblockchain=Ethereum" to set the target chain
to Ethereum.  By default, the blockchain is assumed to be Linea.
Therefore, its only when running the reference tests that this property
needs to be set.  Furthermore, if the reference tests are run without
setting this property ... then an error is reported and the build fails.