fix github action workflows
Since Rust corset cannot be used to check the constraints any longer
(i.e. due to unsupported features being used), it is no longer used in
the `linea-constraints` Makefile.  Thus, we no longer separate the
`zkevm.bin` from the `zkevm.go.bin` (and likewise for the reference
tests).

As a result of all this, the github actions workflow needed to be
updated to remove attempts to run the original corset tool, and also to
default back to using the `zkevm.bin` instead of `zkevm.go.bin` for
`go-corset`.