fix: for CALL_EOA_SUCCESS_WILL_REVERT CallSection
This fixes the case of CALLs of the kind

  scenario/CALL_EOA_SUCCESS_WILL_REVERT

The above was failing when the caller was undergoing deployment. Using
the `setDeploymentInfo(Hub)` method led to using the updated deployment
status (false), and hence seeing the init code as deployed code. This
broke in particular the constraints involving the CODEHASH.