feat: enable coverage report aggregation (#1799)
This puts in place a relatively simple mechanism for acquiring coverage
reports generated from go-corset and aggregating them into a single
file.  To enable coverage report generation, the environment variable
GOCORSET_COVERAGE must be set to the name of the desired output file.
The aggregated coverage generated during the life of the JVM instance
will be written to that file.