feat: use go corset to generate tracejava files (#1858)
This updates all modules and tests within the tracer to use the new
Trace API.  This API has a single Trace class which contains all modules
within as inner classes.

Tidies up the github workflows by removing references to the
original Rust corset tool, as this is no longer used in the pipeline.

Uses the latest version of `go-corset` which supports typed
constants, along with an updated version of the constraints.  This
resolves the outstanding issue with the type of `GAS_LIMIT_MAXIMUM` on
different chains.

This updates `linea-constraints` to the latest commit on master which
includes explicit types for constants `GAS_LIMIT_MAXIMUM` and
`GAS_LIMIT_MINIMUM`.