Tweaking JVM memory config
The challenge with using a high level of parallelism is that we end up
creating a large number of JVMs.  With a minimum heap size of 4GB, we
completely max out memory at 32 processes.  That's ignoring memory
required for the OS and/or go-corset itself.

However, most of these short-lived tests do not need anything like 4GB
to complete.  Setting a minimum heap size of 512MB means the average
test consumes much less RAM, whilst keeping the maximum heap size at 8GB
means there is room for them to expand as needed.

The final challenge with increasing the degree of parallelism is
go-corset itself.  This also uses a large number of cores, even for a
single test.  Therefore, there is a careful balance needed to make
effective use of all available cores.