Enable random sampling by day-of-month
This enables random sampling using the day of the month as a seed for
the RNG.  This means that, in principle, we can recreate any failures by
looking up on which day the test failed.  At the same time, we benefit
from wider testing over a period of time.

A default sample size is given individually for each test, and this can
be overridden using a single environment variable
PRC_CALLTESTS_SAMPLE_SIZE.  Hopefully, this provides the level of
flexibility needed.