feat: tweak weekly tests (#1934)
* sepearate our precompile call tests

Since the precompile call tests have a different dynamic from other
weekly tests, have moved all of them into the group that was previously
the weekly ecpairing suite.  That way, the parallelism for this group
can be dramatically increased without causing other (costly and
unrelated) weekly tests from failing.

* add pull_request trigger

The purpose of this is just for testing to let me run the workflow prior
to it being merged in (since it hasn't been run before).

* Tweaking JVM memory config

The challenge with using a high level of parallelism is that we end up
creating a large number of JVMs.  With a minimum heap size of 4GB, we
completely max out memory at 32 processes.  That's ignoring memory
required for the OS and/or go-corset itself.

However, most of these short-lived tests do not need anything like 4GB
to complete.  Setting a minimum heap size of 512MB means the average
test consumes much less RAM, whilst keeping the maximum heap size at 8GB
means there is room for them to expand as needed.

The final challenge with increasing the degree of parallelism is
go-corset itself.  This also uses a large number of cores, even for a
single test.  Therefore, there is a careful balance needed to make
effective use of all available cores.

* Enable random sampling by day-of-month

This enables random sampling using the day of the month as a seed for
the RNG.  This means that, in principle, we can recreate any failures by
looking up on which day the test failed.  At the same time, we benefit
from wider testing over a period of time.

A default sample size is given individually for each test, and this can
be overridden using a single environment variable
PRC_CALLTESTS_SAMPLE_SIZE.  Hopefully, this provides the level of
flexibility needed.

* consolidate prcCallTests => weeklyTests

This consolidates all of the prcCallTests into the existing weeklyTests
group.  This makes sense as they all have roughly similar test profiles.
The weeklyTests now has a much more aggressive degree of parallelism.

In addition, this sets more suitable default sample sizes for the PRC
calltests.  Specifically, defaults are set to approx 10% of the overall
parameter space.

* split weekly tests into two jobs

This splits the weekly tests up into two distinct jobs so they can run
in parallel.  This hopefully should allow for a larger sampling size.