feat: improve weekly tests ci utilisation (#1955)
* tweak max heap limit for tests

This tweaks the max heap limit to ensure the right value is given when
running on the CI server.

* fix bottleneck with SignedOperationsExtensiveTest

There was a performance bottleneck with the test
SignedOperationsExtensiveTest which was using a MethodOrderer and a
Lifecyle.CLASS attribute.  Both of these meant that the default
parallelism mode of concurrent was being ignored.  That mean this one
test was significantly limiting the amount of parallelism on the CI
machine.

* Enable concurrent execution for specific tests

These tests make use of the `TestInstance.Lifecycle.PER_CLASS` attribute
which prohibit concurrent execution.  To get that, we have to add
another attribute `@Execution(ExecutionMode.CONCURRENT)`

* remove `Lifecycle.PER_CLASS` from test

This removes this annotation from the test class `SixtyThreeSixtyFourthsTests`
 by reworking the sources of the parameterised tests to be static.