feat: update wcp module for assembly implementation
This updates the WCP module in preparation for an assembly
implementation.  In particular, such an implementation only needs the
inputs and outputs in the trace.  In theory, we could even drop the
outputs, but for now they are expected.  Observe that this means every
row in the trace corresponds to a distinct call to WCP --- there are no
"intermediate" rows any more.