Feat/build zkevm bin and traces per fork (#2009)
* update shanghai constraints

Signed-off-by: F Bojarski <ceciestunepoubelle@protonmail.ch>

* ras

Signed-off-by: F Bojarski <ceciestunepoubelle@protonmail.ch>

* update reference tests for shanghai fork

This updates the reference tests to correctly select a ChainConfig
representing Shanghai for the Shanghai tests.  In addition, this updates
the GoCorsetValidator to pass this information when checking
constraints.

This also updates `go-corset` to the latest version which fixes some
issues working with the updated constraints.

* ras

Signed-off-by: F Bojarski <ceciestunepoubelle@protonmail.ch>

* Paris HF

Signed-off-by: F Bojarski <ceciestunepoubelle@protonmail.ch>

* remove shanghai for now on blockchain ref tests

* generate trace files per fork gradle tasks + fix trace imports

* streamline testing with multiple forks

This slightly simplifies the gradle build for testing in the presence of
multiple forks, and also simplifies `CorsetValidator` as well.

* update constraints

* feat: catch e when cast to trace Shanghai

* rebase, update constraints

* feat: refactor nb rows tx max in txndata

* updates `linea-constraints` to latest commit

This updates the linea-constraints submodule to the latest commit on
master.

---------

Signed-off-by: F Bojarski <ceciestunepoubelle@protonmail.ch>
Co-authored-by: F Bojarski <ceciestunepoubelle@protonmail.ch>
Co-authored-by: DavePearce <dave01001110@gmail.com>