support test filter for state test action
This puts in place an optional filter on tests to run for the general
state tests.  This is available when running the action directly via
github, exactly as it already is for the blockchain reference tests.

In addition, this tweaks a few parameters for both the state and
blockchain reference tests.