fix outstanding compilation errors
This fixes the outstanding compilation errors, such that the trace now
compiles and passes at least one test.  However, I believe there are
still problems related to the use of testInfo in static variable
initialisers.  Specifically, we will see that it is null in such
situations, leading to some kind of exception.