Convert to "union of all traces" approach
This converts the trace generation process using go-corset to adopt the
"union of all traces approach" which produces just one Trace interface
which includes *everything*.

At this stage, there remains a problem with BlockData.nROWS_DEPTH which
was being statically imported.  However, this constant differs between
forks and, hence, is not visible in the Trace interface.