fix: ensure loadOpcodes is threadsafe (#2107)
* fix: ensure loadOpcodes is threadsafe

This puts in place some structure within the OpCodes class for ensuring
that loadOpcodes is threadsafe (when previously it wasn't).  This
solution is constrained by the desire to retain the existing design as
much as possible.

* restructure OutOfGasExceptionTest

This restructures one of the tests in OutOfGasExceptionTest to delay
loading opcodes until the fork is known.