fix: ensure loadOpcodes() called by BlockCapturer
This ensures that the BlockCapturer calls `OpCodes.loadOpcodes()` with
the fork before it starts capturing.  Previously, this did not occur and
it resulted in exceptions arising during tracing.

At this stage, there remains an outstanding issue that the fork is
currently fixed on LONDON, and is not determined from BESU.  That is a
separate issue to be fix.