refactor OpCodes API to be fork agnostic
Previously, the OpCodes API was implemented via a number of static
methods and, once a given fork was loaded, it was not possible to load
another fork (doing so would cause a runtime exception).

Therefore, this refactors the OpCodes API to remove entirely the static
API and, instead, replace it with an API based around OpCodes
*instances*.  That is, we have an instance of OpCodes for each fork
under consideration.  This means that, in situations where previously we
could look up OpCodeData information statically, we must now be able to
access a relevant instance of OpCodes.  As such, quite a lot of
refactoring was required to make this possible.

The interface now promoters OpCodeData over OpCode as the primary source
of useful information.  The OpCode class is now largely relegated to
being simply a "mnemonic" with essentially no actual functionality.  In
most cases, this works out well because OpCodeData already had a
mnemonic accessor, meaning whereever we have an instance of OpCodeData
we also have an instance of OpCode.