Use @BeforeAll in conjunction with @BeforeEach
The issue with using @BeforeEach to set the fork and load the opcodes is
that this is not guaranteed to run before methods used as MethodSource's
for parameterised test.  However, @BeforeAll is guaranteed to run before
any test, including any MethodSource methods.  Therefore, splitting up
the process of initialising the TracerTestBase gives the best solution.