preliminary testing
This threads through the testInfo object into the ExecutionEnvironment
such it can be used as part of the test name.  For parametric tests,
this contains information about the parameters used.

At this stage, this only prints out the information for demonstration
purposes.  It does not use it to construct a filename (yet).