remove TestInfoWithChainConfig
This removes the class TestInfoWithChainConfig and, as a result, applies
a large refactoring.  The root cause for this was the presence of the
TestInfo information being stored in TracerTestBase.  It cannot sensibly
be stored at this point when tests are running concurrently.  Instead,
we must obtain the TestInfo object from the test method itself as an
additional parameter.  This issue was impacting the name generation for
parameterised tests based on TestInfo data.  Specifically, the
information was often inaccurate as it was involved in a race condition.

Refactoring the tests to include the TestInfo parameter meant the
benefit of the TestInfoWithChainConfig class was essentially
eliminated.  In particular, to continue using it would require
constructing a TestInfoWithChainConfig instance within each test.
Basically, its easier to remove the class entirely than actually do
that.  In some cases, this leads to a more precise flow of information,
whereby unnecessary TestInfo data is not provided to functions which
don't require it.