feat: switch nightly batch to weekly (#2257)
* switch nightly batch to weekly

This switches one batch of nightly tests to weekly tests.  This also
re-enables a set of nightlys which were previously disabled, though
putting them as weeklys.

* tweak weekly tests CI configuration

Since a number of nightly tests have been moved over to the weekly test
runs, we need to adjust the CI configuration to accomodate this.  In
particular, since the nightly tests represent larger and longer runs.