update EXP asm implementation
This puts in place an update for the EXP asm implementation to account
for the case where CDS < EBS.  In such case, the bytes from CDS upto EBS
need to be filled with zero.

NOTE: this implementation has now been tested against *all* tests in
ExpTest, including those tagged as "nightly".