fix: OutOfGasMemExceptionTest
fixed the "outOfGasExceptionCallDataCopy" test for Prague; transactions
have to clear a floor gas cost

    floorGasCost = 21000 + 10 * wght_byte_count

if we rely on "runOnlyForGasCost" to compute the gas cost of a
transaction we will sometimes run into issues for transactions with
nonempty call data.

Also we can't simply return e.g.

    return max(floorGasCost, toyExecutionEnvironmentV2.runForGasCost());

because various tests expect e.g. an OUT_OF_GAS_EXCEPTION, which may not
be triggered if we inflate the "tx.getGasLimit()" to cover the floor
cost.