feat: integrate stp implementation (#2303)
This updates the tracer to support the assembly implementation of the
STP module.  Mostly this amounts to removing computation done for
intermediate columns in the module, as such computation is now done by
the assembly itself during trace expansion.  Thus, all we need to put
in the trace are the required inputs, along with the expected outputs.