feat: extra tests
ultimately useless: I realized too late that the deployments of system
transactions cannot fail _on Linea_ given that we only process
transactions that are valid. In particular, since the gas limit is fixed
and somewhat large, the only way in which the transaction may fail is if
the sender account has insufficient funds to pay the upfront gas cost.
But that cost includes 'gasLimit * gasPrice' (type 0 tx) and so if it
were to fail, it would be at transaction processing (post Shanghai, that
is).