revert parentBeaconBlockRoot to Optional<String>
This resets the parentBeaconBlockRoot implementation back to what it was
originally, as this is required for two reasons.  Firstly, to support
forks which don't have a parent beacon block root (i.e. those prior to
CANCUN).  Secondly, to support older replays which don't even have an
entry for this field.