embed fork name into replay files
This updates the BlockCapturer so that it embeds the fork name (e.g.
"PRAGUE") into captured replays.  This is not strictly required, but
seems like useful metadata for debugging.  Observe that the fork is not
used within the tracer, it doesn't matter whether it is actually in the
replay or not.