tweak setting for coinbase address
The coinbase address is either determined from the block header, or from
the blocks "extra data".  The latter is done when the Clique consensus
protocol is being used (i.e. all forks prior to PRAGUE).  This fix
simply ensures that old replays will *not* take the coinbase address
from the block header.