Feat/replays for prague with besu and issue1335 (#2458)
* feat add replay for issue 1335

* feat: change timeouts to run replays with besuNode

* revert parentBeaconBlockRoot to Optional<String>

This resets the parentBeaconBlockRoot implementation back to what it was
originally, as this is required for two reasons.  Firstly, to support
forks which don't have a parent beacon block root (i.e. those prior to
CANCUN).  Secondly, to support older replays which don't even have an
entry for this field.

* embed fork name into replay files

This updates the BlockCapturer so that it embeds the fork name (e.g.
"PRAGUE") into captured replays.  This is not strictly required, but
seems like useful metadata for debugging.  Observe that the fork is not
used within the tracer, it doesn't matter whether it is actually in the
replay or not.

* tweak setting for coinbase address

The coinbase address is either determined from the block header, or from
the blocks "extra data".  The latter is done when the Clique consensus
protocol is being used (i.e. all forks prior to PRAGUE).  This fix
simply ensures that old replays will *not* take the coinbase address
from the block header.

---------

Co-authored-by: DavePearce <dave01001110@gmail.com>